refactor: update directory and remove getAPI() temporarily
