--- src/obComponents/obCreateNewNote.ts
@@ -1,1 +1,0 @@
-// TODO Turn a memo into page
